// apps/frontend/packages/api/base/codegen/types/PlanListItem.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

import type { Filial } from "./Filial";
import type { Plan } from "./Plan";
import type { PlanCfoStatusValue2 } from "./PlanCfoStatusValue2";
import type { PlanType } from "./PlanType";
import type { UserSummary } from "./UserSummary";

export type PlanListItem = (Plan & {
    /**
     * @type object
    */
    filial: Filial;
    /**
     * @type object
    */
    planType: PlanType;
    /**
     * @type object
    */
    author: UserSummary;
    /**
     * @description Заполняется только для роли CFO
    */
    myCfoStatus?: (PlanCfoStatusValue2 | null);
});